Major logic refactor - remove all jobparameter classes.
The Designation class will now provide all the inforamtion about what needs to be done to complete it.
Specific designations will inherit from the base Designation class to implement their own logic.

Designations can now provide the ActorStateSatisfyDesignation class with sub-states to carry out and in what order so that designation can be satisfied.

Movement destination class changed to its either a single point or a collection of points.